+++ src/main/java/one/d4d/sessionless/itsdangerous/crypto/Signers.java
@@ -1,5 +1,5 @@
 package one.d4d.sessionless.itsdangerous.crypto;
 
 public enum Signers {
-    DANGEROUS, EXPRESS, OAUTH, TORNADO, RUBY, JWT, UNKNOWN
+    DANGEROUS, EXPRESS, OAUTH, TORNADO, RUBY, JWT, NIMBUSDS, UNKNOWN
 }