+++ src/main/java/one/d4d/sessionless/presenter/KeyPresenter.java
@@ -3,6 +3,7 @@ package one.d4d.sessionless.presenter;
 import burp.config.BurpKeysModelPersistence;
 import burp.config.KeysModel;
 import burp.config.KeysModelListener;
+import com.google.gson.Gson;
 import one.d4d.sessionless.forms.WordlistView;
 import one.d4d.sessionless.forms.dialog.KeyDialog;
 import one.d4d.sessionless.forms.dialog.NewKeyDialog;
@@ -12,7 +13,11 @@ import one.d4d.sessionless.utils.Utils;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -127,6 +132,10 @@ public class KeyPresenter extends Presenter {
         if (d.getKey() != null) {
             model.addKey((SecretKey) d.getKey());
         }
+    }
+
+    public void onButtonLoadDefaultsClick() {
+        Utils.readDefaultSecretKeys("/keys", this.getClass()).forEach(model::addKey);
     }
 
     public void onTableKeysDoubleClick() {